Crate problem044

Problem 44 (Pentagon numbers)

Solution detail

There's not much better here than brute-force. Iterate over pairs of pentagonal numbers, for each one calculating the sum and the difference, and checking if both are pentagonal. Store the difference of any solution, and terminate the loops when the differences grow too large to be a better solution.

The only point of note is that according to Wikipedia, a way to check if a number m is pentagonal is to see if 24m + 1 is a square whose square root is congruent to 5 modulo 6.

Functions

answer
Solve the problem, returning the answer as a String

is_pentagonal
A function to check if a given number is pentagonal.

main

solve
Find the smallest difference between a pair of pentagonal numbers whose sum and difference are also pentagonal.